Move some MultiGet code to new source files
Summary:
Move MultiGet functions to new source files that are included by the
original files.
